backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
firebase-admin==6.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
docker==6.1.3
gitpython==3.1.40
requests==2.31.0
websockets==12.0
aiofiles==23.2.1
pydantic==2.5.0
httpx==0.25.2 